[D2 API] : fix api

+++ DiabloII.Items/DiabloII.Items.Api/Items/Models/ItemProperty.cs
@@ -3,7 +3,7 @@
     public class ItemProperty
     {
         public string Name { get; set; }
-        public int Par { get; set; }
+        public double Par { get; set; }
         public int Minimum { get; set; }
         public int Maximum { get; set; }
         public bool IsPercent { get; set; }
